requirements.txt
requests>=2.25.0
python-decouple>=3.6